+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to FlightTracker are documented here.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.7.0] - 2026-08-16
+- Config backup import/export added to settings
+- Experimental support for 1:32 panels
+- Move "Help" into its own card
+
 ## [v2.6.3] - 2026-08-10
  - Date/Time formatting fixes
  - Numeric seperator adding to config options